% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{add_file_to_resource_path}
\alias{add_file_to_resource_path}
\title{Track data object builder for local csv files}
\usage{
add_file_to_resource_path(file_path = NULL, object = NULL)
}
\arguments{
\item{file_path}{A character. Specify the file_path to the local csv file.}

\item{object}{A gr ranges object.}
}
\value{
list of data specs for a local csv file
}
\description{
Get an object for using local csv to build plots
}
