% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{add_mark}
\alias{add_mark}
\title{Visual marks}
\usage{
add_mark(
  x = NULL,
  xe = NULL,
  x1 = NULL,
  x1e = NULL,
  y = NULL,
  strokeWidth = NULL,
  opacity = NULL,
  row = NULL,
  size = NULL,
  color = NULL,
  stroke = NULL
)
}
\arguments{
\item{stroke}{}
}
\value{
list of mark specifications
}
\description{
Visual marks (e.g., points, lines, and bars) are
the basic graphical elements of a visualization.
}
\details{
For info visit http://gosling-lang.org/docs/mark
}
