% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{arrange_views}
\alias{arrange_views}
\title{Arrange views from compose_view() function.}
\usage{
arrange_views(layout = NULL, views = NULL, ...)
}
\arguments{
\item{layout}{One of "linear" or "circular".}

\item{views}{An object from compose_view() function.}

\item{...}{More options passed to gosling.js.}
}
\description{
Arrange views from compose_view() function.
}
