% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atomic_values_to_list}
\alias{atomic_values_to_list}
\title{atomic_values_to_list}
\usage{
atomic_values_to_list(property_list)
}
\arguments{
\item{property_list}{A character or number or another atomic value.}
}
\value{
List.
}
\description{
atomic_values_to_list
}
