% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleProps.R
\name{brush_styles}
\alias{brush_styles}
\title{Title}
\usage{
brush_styles(
  strokeWidth = NULL,
  strokeOpacity = NULL,
  stroke = NULL,
  opacity = NULL,
  color = NULL
)
}
\arguments{
\item{color}{}
}
\description{
Title
}
