% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_json}
\alias{build_json}
\title{Build gosling spec from R list}
\usage{
build_json(r_list, single_track = TRUE, pretty = TRUE, auto_unbox = TRUE)
}
\arguments{
\item{r_list}{R list object built with other gosling functions}

\item{single_track}{Whether single track. If FALSE then there must be more than 1 track in r_list}

\item{pretty}{Whether to get json with indentation, line breaks etc.}

\item{auto_unbox}{If TRUE will automatically unbox() all atomic vectors of length 1.}
}
\value{
json spec for the gosling output
}
\description{
Build gosling spec from R list
}
