% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component}
\alias{component}
\title{Create react component}
\usage{
component(name)
}
\arguments{
\item{name}{name of the react component}
}
\value{
function to create react element
}
\description{
Create react component
}
