% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{compose_view}
\alias{compose_view}
\title{Compose views from add_single_track() and add_multi_tracks() functions.}
\usage{
compose_view(
  multi = FALSE,
  layout = NULL,
  width = NULL,
  height = NULL,
  centerRadius = NULL,
  tracks,
  ...
)
}
\arguments{
\item{multi}{Whether multiple tracks in the view.}

\item{layout}{One of "linear" or "circular".}

\item{width}{A number interpreted in units of pixel.}

\item{height}{A number interpreted in units of pixel.}

\item{centerRadius}{Specify the proportion of the radius of
the center white space. A number between \link{0,1}, default=0.3}

\item{tracks}{The tracks with add_multi_tracks() function.}

\item{...}{More arguments passed along with view to gosling.js.}
}
\description{
Compose views from add_single_track() and add_multi_tracks() functions.
}
