% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleProps.R
\name{default_track_styles}
\alias{default_track_styles}
\title{Default styles for tracks}
\usage{
default_track_styles(
  textStrokeWidth = NULL,
  textStroke = NULL,
  textFontWeight = NULL,
  textFontSize = NULL,
  textAnchor = NULL,
  select = NULL,
  outlineWidth = NULL,
  outline = NULL,
  mouseOver = NULL,
  matrixExtent = NULL,
  linkStyle = NULL,
  linkMinHeight = NULL,
  linkConnectionType = NULL,
  linePattern = NULL,
  legendTitle = NULL,
  inlineLegend = NULL,
  enableSmoothPath = NULL,
  dy = NULL,
  dx = NULL,
  dashed = NULL,
  curve = NULL,
  brush = NULL,
  backgroundOpacity = NULL,
  background = NULL,
  align = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Default styles for tracks
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel/#style-related-properties
}
