% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleProps.R
\name{event_styles}
\alias{event_styles}
\title{Title}
\usage{
event_styles(
  strokeWidth = NULL,
  strokeOpacity = NULL,
  stroke = NULL,
  opacity = NULL,
  color = NULL,
  arrange = NULL
)
}
\arguments{
\item{arrange}{}
}
\description{
Title
}
