% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pdf.R
\name{export_pdf}
\alias{export_pdf}
\title{Export PDF}
\usage{
export_pdf(
  component_id,
  transparent_background = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{component_id}{A character. The id of the component_id prop passed to the
GoslingComponent function.}

\item{transparent_background}{A Boolean. Determine if the background
should be transparent or not (Default: false).}

\item{session}{A shiny session object.}
}
\value{
None.
}
\description{
Exports PDF
}
