% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_png.R
\name{export_png}
\alias{export_png}
\title{Export PNG}
\usage{
export_png(
  component_id,
  transparent_background = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{component_id}{A character. The id of the component_id prop passed to the
GoslingComponent function.}

\item{transparent_background}{A Boolean. Determine if the background
should be transparent or not (Default: false).}
}
\description{
Exports PNG
}
