% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_file_track_data}
\alias{get_file_track_data}
\title{Track data object builder for local csv files}
\usage{
get_file_track_data(
  file_name,
  chromosomeField = NULL,
  genomicFields = NULL,
  separator = ",",
  sampleLength = 1000,
  headerNames = NULL,
  ...
)
}
\arguments{
\item{file_name}{A character. Specify the file_name.}

\item{chromosomeField}{A character. Specify the name of chromosome data fields.}

\item{genomicFields}{A character vector. Specify the name of genomic data fields.}

\item{separator}{A character. Specify file separator, Default: ','}

\item{sampleLength}{A number. Specify the number of rows loaded from
the URL. Default: 1000}

\item{headerNames}{A character vector. Specify the names of data fields
if a CSV file does not have header row.}

\item{...}{Any other parameters passed to json data object.}
}
\value{
list of data specs for a local csv file
}
\description{
Get an object for using local csv to build plots
}
