% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{gosling_component}
\alias{gosling_component}
\title{Create Gosling component}
\usage{
gosling_component(...)
}
\arguments{
\item{...}{Name of component.}
}
\value{
A function to create the gosling component.
}
\description{
Create Gosling component
}
