% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{GoslingComponent}
\alias{GoslingComponent}
\title{Create Gosling component}
\usage{
GoslingComponent(...)
}
\arguments{
\item{...}{Name of component.}
}
\value{
A function to create the gosling component.
}
\description{
Create Gosling component
}
