% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{goslingOutput}
\alias{goslingOutput}
\title{gosling output function}
\usage{
goslingOutput(outputId)
}
\arguments{
\item{outputId}{ID of the output element}
}
\value{
reactOutput HTML for UI render
}
\description{
gosling output function for shiny use.
Must use this function instead of shiny output functions.
}
