% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_atomic_field}
\alias{is_atomic_field}
\title{is_atomic_field}
\usage{
is_atomic_field(field_name)
}
\arguments{
\item{field_name}{A character or number or another atomic value.}
}
\value{
List.
}
\description{
is_atomic_field
}
