% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.gosling}
\alias{print.gosling}
\title{Print method for the gosling component}
\usage{
\method{print}{gosling}(x, ...)
}
\arguments{
\item{x}{A gosling object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
r list without NULL values
}
\description{
Print method for the gosling component
}
