% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderGosling}
\alias{renderGosling}
\title{gosling render function}
\usage{
renderGosling(expr)
}
\arguments{
\item{expr}{The code for rendering gosling plot.
Must be a gosling object built with shiny.gosling::gosling()}
}
\value{
A function which can be assigned to an output in a Shiny server function.
}
\description{
gosling render function for shiny use
}
