% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{track_data_transform}
\alias{track_data_transform}
\title{Data transformer}
\usage{
track_data_transform(type = NULL, field = NULL, oneOf = NULL, not = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{
list of data transformations specs
}
\description{
Do data transformations
}
\details{
For info visit http://gosling-lang.org/docs/data#data-transform
}
