% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{track_data_transforms}
\alias{track_data_transforms}
\title{Combine multiple data transforms}
\usage{
track_data_transforms(...)
}
\arguments{
\item{...}{Multiple data transform specs separated by comma.}
}
\value{
list of multiple data transform specs
}
\description{
Combine multiple data transforms
}
