% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_gosling}
\alias{use_gosling}
\title{Initiate gosling}
\usage{
use_gosling()
}
\value{
Gosling initiator HTML.
}
\description{
Add this function at the begining of ui. This is needed for
gosling to work in shiny plots.
}
