% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel}
\alias{visual_channel}
\title{Generic visual channel builder}
\usage{
visual_channel(field = NULL, type = NULL, range = NULL, domain = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Generic visual channel builder
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#encode-a-visual-channel
}
