% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_color}
\alias{visual_channel_color}
\title{color visual channel}
\usage{
visual_channel_color(
  field = NULL,
  title = NULL,
  type = NULL,
  scaleOffset = NULL,
  scale = NULL,
  legend = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
color visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#color
}
