% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_opacity}
\alias{visual_channel_opacity}
\title{opacity visual channel}
\usage{
visual_channel_opacity(
  field = NULL,
  type = NULL,
  range = NULL,
  domain = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
opacity visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#opacity
}
