% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_size}
\alias{visual_channel_size}
\title{size visual channel}
\usage{
visual_channel_size(
  field = NULL,
  type = NULL,
  range = NULL,
  domain = NULL,
  ...
)
}
\arguments{
\item{field}{A character. Name of the data field.}

\item{type}{A character. Must be "genomic". Specify the data type.}

\item{range}{A vector of characters or numbers. Values of the visual
channel. Range to be specified like \code{range = c(min_size, max_size)}}

\item{domain}{A vector of characters or numbers. Values of the data.}

\item{...}{Any other parameters to pass to gosling.js.}
}
\value{
List object.
}
\description{
size visual channel
}
\details{
For more info visit
http://gosling-lang.org/docs/visual-channel#size
}
