% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_stroke}
\alias{visual_channel_stroke}
\title{stroke visual channel}
\usage{
visual_channel_stroke(
  field = NULL,
  title = NULL,
  type = NULL,
  scaleOffset = NULL,
  legend = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
stroke visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#stroke
}
