% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_stroke_width}
\alias{visual_channel_stroke_width}
\title{stroke width visual channel}
\usage{
visual_channel_stroke_width(
  field = NULL,
  type = NULL,
  range = NULL,
  domain = NULL,
  ...
)
}
\arguments{
\item{field}{A character. Name of the data field.}

\item{type}{A character. Must be "genomic". Specify the data type.}

\item{range}{A vector of characters or numbers. Values of the visual channel.}

\item{domain}{A vector of characters or numbers. Values of the data.}

\item{...}{Any other parameters to pass to gosling.js.}
}
\value{
List object.
}
\description{
stroke width visual channel
}
\details{
For more info visit
http://gosling-lang.org/docs/visual-channel#strokewidth
}
