% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_text}
\alias{visual_channel_text}
\title{text visual channel}
\usage{
visual_channel_text(
  field = NULL,
  type = NULL,
  range = NULL,
  domain = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
text visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#text
}
