% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_tooltip}
\alias{visual_channel_tooltip}
\title{tooltip visual channel}
\usage{
visual_channel_tooltip(field = NULL, type = NULL, alt = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{
list object with tooltip list object
}
\description{
tooltip visual channel
}
\details{
For more info visit https://gosling.js.org/ and check for tooltip implementation
}
