% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_tooltips}
\alias{visual_channel_tooltips}
\title{Combine tooltips into a list}
\usage{
visual_channel_tooltips(...)
}
\arguments{
\item{...}{}
}
\value{
json list with tooltips combined into a single spec
}
\description{
Combine tooltips into a list
}
