% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_x}
\alias{visual_channel_x}
\title{x and xe axis visual channel}
\usage{
visual_channel_x(
  field = NULL,
  type = NULL,
  legend = NULL,
  grid = NULL,
  axis = NULL,
  aggregate = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
x and xe axis visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#x--xe
}
