% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_to_extent.R
\name{zoom_to_extent}
\alias{zoom_to_extent}
\title{Zoom to extent}
\usage{
zoom_to_extent(
  component_id,
  view_id,
  duration = 1000,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{component_id}{A character. The id of the component_id prop passed to the
GoslingComponent function.}

\item{view_id}{A character. The ID of a view that you want to control. This ID
is consistent to what you specify as track.id in your spec.}

\item{duration}{A numeric. A duration of the animated transition in ms
(Default: 1000).}

\item{session}{A shiny session object.}
}
\value{
None.
}
\description{
Zooms out to see the entire view_id passed to this function.
}
