% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_to_gene.R
\name{zoom_to_gene}
\alias{zoom_to_gene}
\title{Zoom to gene}
\usage{
zoom_to_gene(
  component_id,
  view_id,
  gene,
  padding = 0,
  duration = 1000,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{component_id}{A character. The id of the component_id prop passed to the
GoslingComponent function.}

\item{view_id}{A character. The ID of a view that you want to control. This ID
is consistent to what you specify as track.id in your spec.}

\item{gene}{A character. The target gene symbol that your view
should be navigated to (e.g., MYC).}

\item{padding}{A numeric. This determines the padding around the specified position.
The unit of this number is a base pair (Default: 0).}

\item{duration}{A numeric. A duration of the animated transition in ms
(Default: 1000).}

\item{session}{A shiny session object.}
}
\description{
This function makes a view navigate to a certain gene with animated transition.
}
