% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{add_mark}
\alias{add_mark}
\title{Visual marks}
\usage{
add_mark(
  x = NULL,
  xe = NULL,
  x1 = NULL,
  x1e = NULL,
  y = NULL,
  strokeWidth = NULL,
  opacity = NULL,
  row = NULL,
  size = NULL,
  color = NULL,
  stroke = NULL
)
}
\arguments{
\item{x}{An object returned by visual_channel_x().}

\item{xe}{An object returned by visual_channel_x().}

\item{x1}{An object returned by visual_channel_x().}

\item{x1e}{An object returned by visual_channel_x().}

\item{y}{An object returned by visual_channel_y().}

\item{strokeWidth}{A number or an object returned by visual_channel_stroke_width().}

\item{opacity}{A number or an object returned by visual_channel_opacity().}

\item{row}{A factor data column Channel row is used with channel y to stratify a visualization with categorical values.}

\item{size}{A number or an object returned by visual_channel_size().}

\item{color}{A character or an object returned by visual_channel_color().}

\item{stroke}{A number or an object returned by visual_channel_stroke().}
}
\value{
list of mark specifications
}
\description{
Visual marks (e.g., points, lines, and bars) are
the basic graphical elements of a visualization.
}
\details{
For info visit http://gosling-lang.org/docs/mark
}
