% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{add_mark}
\alias{add_mark}
\title{Title}
\usage{
add_mark(
  x = NULL,
  xe = NULL,
  x1 = NULL,
  x1e = NULL,
  y = NULL,
  strokeWidth = NULL,
  opacity = NULL,
  row = NULL,
  size = NULL,
  color = NULL,
  stroke = NULL
)
}
\arguments{
\item{stroke}{}
}
\description{
Title
}
