% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{add_single_track}
\alias{add_single_track}
\title{Add a single track to the plot of a mark type ( plot type )}
\usage{
add_single_track(
  id = NULL,
  data = NULL,
  mark = NULL,
  assembly = NULL,
  row = NULL,
  size = NULL,
  color = NULL,
  strokeWidth = NULL,
  opacity = NULL,
  x = NULL,
  xe = NULL,
  x1 = NULL,
  x1e = NULL,
  y = NULL,
  stroke = NULL,
  width = NULL,
  height = NULL,
  dataTransform = NULL,
  ...
)
}
\arguments{
\item{id}{Optional argument to assign an id to the track.}

\item{data}{An object of from track_data() function.}

\item{mark}{Type of plot. One of c("point", "line", "rect", "bar", "area",
"link", "triangle", "text"). Each mark type has some supported visual channel.

Different marks support different visual channels:
\itemize{
\item point:    x, y, row, size, color, strokeWidth, opacity
\item line:    x, y, row, color, strokeWidth
\item rect:    x, xe, row, color, strokeWidth, opacity
\item bar:    x, y, row, color, strokeWidth, opacity
\item area:    x, y, row, color, strokeWidth
\item link:    x, xe, x1, x1e, color, opacity
\item triangle:    x, xe, row, size, color, opacity
\item text:    x, xe, row, color, opacity
}}

\item{assembly}{Currently support "hg38", "hg19", "hg18",
"hg17", "hg16", "mm10", "mm9". Defaults to "hg38".}

\item{row}{An object of from visual_channel_row().}

\item{size}{An object of from visual_channel_size() OR an atomic number.}

\item{color}{An object of from visual_channel_color() OR and atomic
character hex code of the form "#123456".}

\item{strokeWidth}{An object of from visual_channel_stroke_width() OR an
atomic number.}

\item{opacity}{An object of from visual_channel_opacity() OR and atomic
ratio from 0 to 1.}

\item{x}{An object of from visual_channel_x() OR an atomic value.}

\item{xe}{An object of from visual_channel_x() OR an atomic value.}

\item{x1}{An object of from visual_channel_x() OR an atomic value.}

\item{x1e}{An object of from visual_channel_x() OR an atomic value.}

\item{y}{An object of from visual_channel_y() OR an atomic value.}

\item{stroke}{An object of from visual_channel_stroke() function OR a
character of hex color code like "#123456".}

\item{width}{A number interpreted in units of pixel.}

\item{height}{A number interpreted in units of pixel.}

\item{dataTransform}{An object of from track_data_transform() function.}

\item{...}{Any other arguments to be passed onto gosling.js.}
}
\description{
This function constructs a single track from the inputs. The inputs
can be id, data, mark etc. Please check gosling.js documentation for
advabced usage.
}
