% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{arrange_views}
\alias{arrange_views}
\title{Arrange views}
\usage{
arrange_views(layout = NULL, views = NULL, listify = TRUE, ...)
}
\arguments{
\item{layout}{One of "linear" or "circular".}

\item{views}{An object from compose_view() function.}

\item{listify}{A Boolean. Convert views to list..}

\item{...}{More options passed to gosling.js.}
}
\value{
list object.
}
\description{
Arrange views from compose_view() function.
}
