% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleProps.R
\name{brush_styles}
\alias{brush_styles}
\title{style of the brush mark}
\usage{
brush_styles(
  strokeWidth = NULL,
  strokeOpacity = NULL,
  stroke = NULL,
  opacity = NULL,
  color = NULL
)
}
\arguments{
\item{color}{}
}
\description{
Customize the style of the brush mark in the \code{rangeSelect} mouse event.
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel/#type-brush
}
