% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleProps.R
\name{brush_styles}
\alias{brush_styles}
\title{style of the brush mark}
\usage{
brush_styles(
  strokeWidth = NULL,
  strokeOpacity = NULL,
  stroke = NULL,
  opacity = NULL,
  color = NULL
)
}
\arguments{
\item{strokeWidth}{A number. stroke width of the marks when mouse events are triggered.}

\item{strokeOpacity}{A number.}

\item{stroke}{A character. Stroke color of the marks when mouse events are triggered.}

\item{opacity}{A number. Opacity of the marks when mouse events are triggered.}

\item{color}{A character. Color of the marks when mouse events are triggered.}
}
\value{
List object with brush styles.
}
\description{
Customize the style of the brush mark in the \code{rangeSelect} mouse event.
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel/#type-brush
}
