% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleProps.R
\name{event_styles}
\alias{event_styles}
\title{Mouse event styles}
\usage{
event_styles(
  strokeWidth = NULL,
  strokeOpacity = NULL,
  stroke = NULL,
  opacity = NULL,
  color = NULL,
  arrange = NULL
)
}
\arguments{
\item{arrange}{}
}
\value{
list of event styles
}
\description{
The styles defined here will be applied to the targets of mouse events, such as a point mark after user click mouse.
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel/#type-eventstyle
}
