% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleProps.R
\name{event_styles}
\alias{event_styles}
\title{Mouse event styles}
\usage{
event_styles(
  strokeWidth = NULL,
  strokeOpacity = NULL,
  stroke = NULL,
  opacity = NULL,
  color = NULL,
  arrange = NULL
)
}
\arguments{
\item{strokeWidth}{A number. stroke width of the marks when mouse events are triggered.}

\item{strokeOpacity}{A number.}

\item{stroke}{A character. Stroke color of the marks when mouse events are triggered.}

\item{opacity}{A number. Opacity of the marks when mouse events are triggered.}

\item{color}{A character. Color of the marks when mouse events are triggered.}

\item{arrange}{A character. One of "behind", "front". Show event effects behind or in front of marks.}
}
\value{
List object with event styles.
}
\description{
The styles defined here will be applied to the targets of mouse events, such as a point mark after user click mouse.
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel/#type-eventstyle
}
