% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gosling}
\alias{gosling}
\title{Build gosling plot object}
\usage{
gosling(component_id, composed_views, clean_braces = TRUE)
}
\arguments{
\item{component_id}{Assign a component id to use other api like zoom.}

\item{composed_views}{The views composed with arrange_views.}

\item{single_track}{Whether single track view.}
}
\value{
Gosling component for rendering on R shiny apps
}
\description{
Build gosling plot object
}
