% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{track_data}
\alias{track_data}
\title{Data object builder}
\usage{
track_data(
  url,
  type,
  separator = NULL,
  sampleLength = NULL,
  headerNames = NULL,
  genomicFields = NULL,
  chromosomeField = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
list of data specs
}
\description{
Build the data object for golsing plots
}
\details{
For info visit http://gosling-lang.org/docs/data
}
