% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{track_data_csv}
\alias{track_data_csv}
\title{Data object builder for a csv file}
\usage{
track_data_csv(
  url,
  separator = ",",
  sampleLength = 1000,
  headerNames = NULL,
  genomicFields = NULL,
  chromosomeField = NULL
)
}
\arguments{
\item{url}{A character. Specify the URL address of the data file.}

\item{separator}{A character. Specify file separator, Default: ','}

\item{sampleLength}{A number. Specify the number of rows loaded from
the URL. Default: 1000}

\item{headerNames}{A character vector. Specify the names of data fields
if a CSV file does not have header row.}

\item{genomicFields}{A character vector. Specify the name of genomic
data fields.}

\item{chromosomeField}{A character. Specify the name of chromosome
data fields.}

\item{...}{Any other parameters passed to json data object.}
}
\value{
list of data specs for a csv file
}
\description{
Build the data object for gosling plots
}
