% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{track_data_gr}
\alias{track_data_gr}
\title{Data object builder for a GRanges object by locally saving it}
\usage{
track_data_gr(granges)
}
\arguments{
\item{granges}{A GRanges object from the GenomicRanges package with seqnames and ranges}
}
\value{
list of data specs for a csv file
}
\description{
Build the data object for gosling plots
}
