% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel}
\alias{visual_channel}
\title{Generic visual channel builder}
\usage{
visual_channel(field = NULL, type = NULL, range = NULL, domain = NULL, ...)
}
\arguments{
\item{field}{A character. Name of the data field.}

\item{type}{A character. Must be "genomic". Specify the data type.}

\item{range}{A vector of characters or numbers. Values of the visual channel.}

\item{domain}{A vector of characters or numbers. Values of the data.}

\item{...}{Any other parameters to pass to gosling.js.}
}
\value{
List object.
}
\description{
Generic visual channel builder
}
\details{
For more info visit
http://gosling-lang.org/docs/visual-channel#encode-a-visual-channel
}
