% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_row}
\alias{visual_channel_row}
\title{row visual channel}
\usage{
visual_channel_row(
  field = NULL,
  type = NULL,
  padding = NULL,
  legend = NULL,
  grid = NULL,
  clip = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
row visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#row
}
