% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_size}
\alias{visual_channel_size}
\title{Range to be specified like \code{range = c(min_size, max_size)}}
\usage{
visual_channel_size(
  field = NULL,
  type = NULL,
  range = NULL,
  domain = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Range to be specified like \code{range = c(min_size, max_size)}
}
