% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_size}
\alias{visual_channel_size}
\title{size visual channel}
\usage{
visual_channel_size(
  field = NULL,
  type = NULL,
  range = NULL,
  domain = NULL,
  ...
)
}
\arguments{
\item{range}{Range to be specified like \code{range = c(min_size, max_size)}}

\item{...}{}
}
\description{
size visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#size
}
