% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_stroke_width}
\alias{visual_channel_stroke_width}
\title{stroke width visual channel}
\usage{
visual_channel_stroke_width(
  field = NULL,
  type = NULL,
  range = NULL,
  domain = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
stroke width visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#strokewidth
}
