% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_tooltip}
\alias{visual_channel_tooltip}
\title{tooltip visual channel}
\usage{
visual_channel_tooltip(field = NULL, type = NULL, alt = NULL, ...)
}
\arguments{
\item{field}{A character. Name of the data field.}

\item{type}{A character. Must be "genomic". Specify the data type.}

\item{alt}{A character.}

\item{...}{Any other parameters to pass to gosling.js.}
}
\value{
List object. list object with tooltip list object
}
\description{
tooltip visual channel
}
\details{
For more info visit https://gosling.js.org/ and check for tooltip implementation
}
