% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualChannels.R
\name{visual_channel_y}
\alias{visual_channel_y}
\title{y and ye axis visual channel}
\usage{
visual_channel_y(
  field = NULL,
  zeroBaseline = NULL,
  type = NULL,
  legend = NULL,
  grid = NULL,
  flip = NULL,
  baseline = NULL,
  axis = NULL,
  aggregate = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
y and ye axis visual channel
}
\details{
For more info visit http://gosling-lang.org/docs/visual-channel#y--ye
}
